+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * JSONUtil增加isJson、isJsonObj、isJsonArray方法（issue#IGT7A@Gitee）
 * Week增加方法toChinese（感谢@【北京】宁静）
 * AbstractCaptcha增加方法generateCode，用于抽象验证码字符串生成，提供可定制验证码字符串生成（感谢@【杭州】KIWI @【杭州】t-io ）
+* NetUtil增加netCat方法
 
 ### Bug修复
 * 修复ExcelUtil.readBySax方法判断03或07格式导致的问题（issue#IGT7X@Gitee）